\name{Curvedata}
\alias{Curvedata}
\docType{data}
\title{
Curvedata is used for exemplifying the usage of function calED()
}
\description{
Four sets of growth curves and correlated standardlized natural signal values from Duller (2007).
}
\usage{data(Curvedata)}
\format{
 A list contains four sets of growth curves and correlated standardlized natural signal values
}
\references{
Duller, G.A.T., 2007. Assessing the error on equivalent dose estimates derived from single aliquot regenerative dose measurements. Ancient TL 25(1), pp. 15-24.
}
\examples{
data(Curvedata)
print(Curvedata)
}
\keyword{growth curve}
