\name{print.RadialPlotter}
\alias{print.RadialPlotter}
\title{
Print object "RadialPlotter"
}
\description{
Print object "RadialPlotter"
}
\usage{
\method{print}{RadialPlotter}(x, ...)
}
\arguments{
  \item{x}{An object of class "RadialPlotter"
}
  \item{\dots}{do not use
}
}
\value{
Error message, parameters, BIC and maximum logged likelihood value
}
\seealso{
 \link{RadialPlotter}
}
\keyword{OSL dating}
