\name{EDdata}
\alias{EDdata}
\docType{data}
\title{
Equivalent dose values
}
\description{
Two sets of equivalent dose values.
}
\usage{data(EDdata)}
\format{
A list that contains two sets of equivalent dose values:
\describe{
  \item{gl11}{35 equivalent dose values of a sand sample from the Tengger Desert (Peng and Han, 2013)}
  \item{al3}{84 equivalent dose values of an alluvial deposit from the andean precordillera (Schmidt et al., 2012)}
}
}
\references{
Peng J, Han FQ, 2013. Selections of fast-component OSL signal using sediments from the south edge of Tengger Desert. Acta Geoscientica Sinica, 34: 757-762.

Schmidt S, Tsukamoto S, Salomon E, Frechen M, Hetzel R, 2012. Optical dating of alluvial deposits at the orogenic front of the andean precordillera (Mendoza, Argentina). Geochronometria, 39: 62-75.
}
\examples{
  data(EDdata)
  names(EDdata)
}
\keyword{OSL dating}
\keyword{equivalent dose}
