\name{numOSL-package}
\alias{numOSL-package}
\docType{package}
\title{Package for tackling basic numeric problems in optically stimulated luminescence dating}
\description{
Package for routine numeric optimization and data visualization in optically stimulated luminescence dating.
}
\details{
\tabular{ll}{
Package: \tab numOSL\cr
Type: \tab Package\cr
Version: \tab 2.1\cr
Date: \tab 2017-01-22\cr
License: \tab GPL-3\cr
}
}
\author{
\tabular{ll}{
Jun Peng \tab University of Chinese Academy of Sciences, Beijing, China \cr
Bo Li \tab University of Wollongong, Wollongong, Australia
}

\bold{Package maintainer}\cr
Jun Peng
\email{pengjun10@mails.ucas.ac.cn}

\bold{Related package projects}\cr
R program KMS
\url{https://github.com/pengjunUCAS/KMS}\cr
R package tgcd
\url{https://CRAN.R-project.org/package=tgcd}
}
\references{
Peng J, Dong ZB, Han FQ, Long H, Liu XJ, 2013. R package numOSL: numeric routines for optically stimulated luminescence dating. Ancient TL, 31(2): 41-48.
}
\keyword{OSL dating}
\keyword{BIN file}
\keyword{statistical age models}
\keyword{decay curve}
\keyword{growth curve}
\keyword{equivalent dose}
\keyword{dose rate}
