\name{numOSL-package}
\alias{numOSL-package}
\docType{package}
\title{Package for tackling basic numeric problems in optically stimulated luminescence dating}
\description{
Package for routine numeric optimization and data visualization in optically stimulated luminescence dating.
}
\details{
\tabular{ll}{
Package: \tab numOSL\cr
Type: \tab Package\cr
Version: \tab 2.7\cr
Date: \tab 2023-09-17\cr
License: \tab GPL-3\cr
}
}
\author{
\tabular{ll}{
Jun Peng \tab Hunan University of Science and Technology (HNUST), Xiangtan, China \cr
Bo Li \tab University of Wollongong (UOW), Wollongong, Australia\cr
Chunxin Wang \tab University of Science and Technology of China, Hefei, China
}

\bold{Package maintainer}\cr
Jun Peng
\email{pengjun10@mails.ucas.ac.cn}

\bold{Related package projects}\cr
R program KMS
\url{https://github.com/pengjunUCAS/KMS}\cr
R package tgcd
\url{https://CRAN.R-project.org/package=tgcd}
}
\references{
Peng J, Dong ZB, Han FQ, Long H, Liu XJ, 2013. R package numOSL: numeric routines for optically stimulated luminescence dating. Ancient TL, 31(2): 41-48.

Peng J, Li Bo, 2017. Single-aliquot Regenerative-Dose (SAR) and Standardised Growth Curve (SGC) Equivalent Dose Determination in a Batch Model Using the R Package 'numOSL'. Ancient TL, 35(2): 32-53.
}
\keyword{OSL dating}
\keyword{BIN file}
\keyword{decay curve}
\keyword{growth curve}
\keyword{equivalent dose}
\keyword{batch analysis}
\keyword{statistical age models}
\keyword{dose rate}

