\name{numOSL-package}
\alias{numOSL-package}
\docType{package}
\title{Package for tackling basic numeric problems in optically stimulated luminescence dating}
\description{
Package for routine numeric optimization and data visualization for optically stimulated luminescence dating.
}
\details{
\tabular{ll}{
Package: \tab numOSL\cr
Type: \tab Package\cr
Version: \tab 1.8\cr
Date: \tab 2015-05-05\cr
License: \tab GPL-3\cr
}
}
\author{
\tabular{ll}{
Peng Jun \tab University of Chinese Academy of Sciences, Beijing, China
}
\bold{Package maintainer}

Peng Jun
\email{pengjun10@mails.ucas.ac.cn}
}
\references{
Peng J, Dong ZB, Han FQ, Long H, Liu XJ, 2013. R package numOSL: numeric routines for optically stimulated luminescence dating. Ancient TL, 31(2): 41-48.
}
\keyword{OSL dating}
\keyword{BIN file}
\keyword{statistical age models}
\keyword{OSL decay curve}
\keyword{fast-component equivalent dose}
\keyword{equivalent dose}
\keyword{annual dose rate}
