\name{print.RadialPlotter}
\alias{print.RadialPlotter}
\title{
Print an object of class "RadialPlotter"
}
\description{
Print an object of class "RadialPlotter".
}
\usage{
\method{print}{RadialPlotter}(x, ...)
}
\arguments{
  \item{x}{An object of class "RadialPlotter"}
  \item{\dots}{do not use}
}
\value{
An error message, the estimates, the associated BIC and maximum logged likelihood value.
}
\seealso{
 \link{RadialPlotter}
}
\keyword{Radial Plot}
\keyword{statistical age models}
